clean up classnames, add message to map snackbar
